apps/api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
firebase-admin==6.2.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
qrcode[pil]==7.4.2
requests==2.31.0
user-agents==2.2.0
python-dotenv==1.0.0
sendgrid==6.10.0
google-api-python-client==2.108.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.1.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2